后台开发路线分享，并完善菜单结构

+++ src/main/java/com/project/platform/entity/ScenicInfo.java
@@ -24,7 +24,7 @@ public class ScenicInfo  {
      */
     private String name;
 
-    private float  originalPrice;
+    private BigDecimal originalPrice;
     /**
      * 收费价格
      */
